PR #722: Split stub generation tests
